% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.list.r
\name{get.list}
\alias{get.list}
\title{Create regular data frames from regions - industries matrices}
\usage{
get.list(mat)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns (or the other way around)}
}
\description{
This function creates regular data frames with three columns (regions, industries, count) from (incidence) matrices (wide to long format) using the reshape2 package
}
\examples{
## generate a region - industry data frame
set.seed(31)
region <- c("R1", "R1", "R1", "R1", "R2", "R2", "R3", "R4", "R5")
industry <- c("I1", "I2", "I3", "I4", "I1", "I2", "I1", "I1", "I3")
data <- data.frame (region, industry)
data$count <- 1

## run the function
get.list (mat)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\seealso{
\code{\link{get.matrix}}
}
\keyword{adjacency}
\keyword{economic}
\keyword{economics,}
\keyword{edge}
\keyword{geography}
\keyword{geography,}
\keyword{lists,}
\keyword{matrix,}
\keyword{package,}
\keyword{reshape}

