% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hoover.Gini.r
\name{Hoover.Gini}
\alias{Hoover.Gini}
\title{Compute the Hoover Gini from regional counts in population and industry}
\usage{
Hoover.Gini(ind, pop)
}
\arguments{
\item{ind}{A vector of industrial regional count}

\item{pop}{A vector of population regional count}
}
\description{
This function computes the Hoover Gini from regional shares in population and industry
}
\examples{
## generate vectors of industrial and population count
ind <- c(0, 10, 10, 30, 50)
pop <- c(5, 10, 10, 25, 30)

## run the function
Hoover.Gini (ind, pop)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Hoover, E.M. (1936) The Measurement of Industrial Localization, \emph{The Review of Economics and Statistics} \strong{18} (1): 162-171
}
\seealso{
\code{\link{location.quotient}}, \code{\link{locational.Gini}}
}
\keyword{concentration}
\keyword{inequality}

