% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.quotient.r
\name{location.quotient}
\alias{location.quotient}
\title{Compute location quotients from regions - industries matrices}
\usage{
location.quotient(mat, binary = FALSE)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns}

\item{binary}{Logical; shall the returned output be a dichotomized version (0/1) of the location quotient? Defaults to FALSE (the full values of the location quotient will be returned), but can be set to TRUE (location quotient values above 1 will be set to 1 & location quotient values below 1 will be set to 0)}
}
\description{
This function computes location quotients from (incidence) regions - industries matrices. The numerator is the share of a given industry in a given region. The denominator is the share of a this industry in a larger economy (overall country for instance). This index is also refered to as the index of Revealed Comparative Advantage (RCA) following Ballasa (1965), or the Hoover-Balassa index.
}
\examples{
## generate a region - industry matrix
set.seed(31)
mat <- matrix(sample(0:100,20,replace=T), ncol = 4)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function
location.quotient (mat)
location.quotient (mat, binary = TRUE)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Balassa, B. (1965) Trade Liberalization and Revealed Comparative Advantage, \emph{The Manchester School} \strong{33}: 99-123.
}
\seealso{
\code{\link{RCA}}
}
\keyword{concentration}
\keyword{specialization}

