% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locational.Gini.curve.r
\name{locational.Gini.curve}
\alias{locational.Gini.curve}
\title{Plot a locational Gini curve from regions - industries matrices}
\usage{
locational.Gini.curve(mat, col = 1)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns}

\item{col}{For which industry shall the curve be plotted? Indicate the correponding column number}
}
\description{
This function plots a locational Gini curve following Krugman from regions - industries matrices.
}
\examples{
## generate a region - industry matrix
set.seed(31)
mat <- matrix(sample(0:100,20,replace=T), ncol = 4)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function
locational.Gini.curve (mat, col = 1)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Krugman P. (1991) \emph{Geography and Trade}, MIT Press, Cambridge (chapter 2 - p.56)
}
\seealso{
\code{\link{locational.Gini}}
}
\keyword{concentration}
\keyword{inequality}

