% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz.coord.r
\name{Lorenz.coord}
\alias{Lorenz.coord}
\title{Computes the coordinates used to plot a Lorenz curve from regional industrial counts}
\usage{
Lorenz.coord(mat)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns. The input can also be a vector of industrial regional count (a matrix with n regions in rows and a single column).}
}
\description{
This function computes the coordinates used to plot a Lorenz curve from regional industrial counts. This curve gives an indication of the unequal distribution of an industry accross regions.
}
\examples{
## generate vectors of industrial count
ind <- c(0, 10, 10, 30, 50)

## run the function
Lorenz.coord (ind)

## plot the coordinates returned by the function to draw a Lorenz curve
plot (Lorenz.coord (ind)$cum.reg, Lorenz.coord (ind)$cum.out,
type = "l", main = "Lorenz curve",
xlab="Cumulative proportion of regions",
ylab="Cumulative proportion of industrial output",
xlim=c(0, 1), ylim=c(0, 1))
abline (0,1, col = "red")

## generate a region - industry matrix
mat = matrix (
c (0, 1, 0, 0,
0, 1, 0, 0,
0, 1, 0, 0,
0, 1, 0, 1,
0, 1, 1, 1), ncol = 4, byrow = T)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function by aggregating all industries
Lorenz.coord (rowSums(mat))

## run the function for industry #1 only (perfect equality)
Lorenz.coord (mat[,1])

## run the function for industry #2 only (perfect equality)
Lorenz.coord (mat[,2])

## run the function for industry #3 only (perfect unequality)
Lorenz.coord (mat[,3])

## run the function for industry #4 only (top 40\% produces 100\% of the output)
Lorenz.coord (mat[,4])

}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Lorenz, M. O. (1905) Methods of measuring the concentration of wealth, \emph{Publications of the American Statistical Association} \strong{9}: 209–219
}
\seealso{
\code{\link{Hoover.Gini}}, \code{\link{locational.Gini}}, \code{\link{locational.Gini.curve}}, \code{\link{Hoover.curve}}, \code{\link{Gini}}
}
\keyword{concentration}
\keyword{inequality}

