% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz.curve.r
\name{Lorenz.curve}
\alias{Lorenz.curve}
\title{Plot a Lorenz curve from regional industrial counts}
\usage{
Lorenz.curve(ind)
}
\arguments{
\item{ind}{A vector of industrial regional count}
}
\description{
This function plots a Lorenz curve from regional industrial counts. This curve gives an indication of the unequal distribution of an industry accross regions.
}
\examples{
## generate vectors of industrial count
ind <- c(0, 10, 10, 30, 50)

## run the function
Lorenz.curve (ind)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Lorenz, M. O. (1905) Methods of measuring the concentration of wealth, \emph{Publications of the American Statistical Association} \strong{9}: 209–219
}
\seealso{
\code{\link{Hoover.curve}}, \code{\link{Gini}}
}
\keyword{concentration}
\keyword{inequality}

