% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rca.r
\name{rca}
\alias{rca}
\title{Compute an index of revealed comparative advantage (RCA) from regions - industries matrices}
\usage{
rca(mat, binary = FALSE)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns}

\item{binary}{Logical; shall the returned output be a dichotomized version (0/1) of the RCA? Defaults to FALSE (the full values of the RCA will be returned), but can be set to TRUE (RCA above 1 will be set to 1 & RCA values below 1 will be set to 0)}
}
\value{
A matrix representing the index of revealed comparative advantage (RCA) or location quotient. Each cell in the matrix corresponds to the RCA value for a specific region and industry. If the `binary` parameter is set to TRUE, the returned matrix will be dichotomized, with values above 1 set to 1 and values below 1 set to 0.
}
\description{
This function computes an index of revealed comparative advantage (RCA) from (incidence) regions - industries matrices. The numerator is the share of a given industry in a given region. The denominator is the share of a this industry in a larger economy (overall country for instance). This index is also refered to as a location quotient, or the Hoover-Balassa index.
}
\examples{
## generate a region - industry matrix
set.seed(31)
mat <- matrix(sample(0:100, 20, replace = TRUE), ncol = 4)
rownames(mat) <- c("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c("I1", "I2", "I3", "I4")

## run the function
rca(mat)
rca(mat, binary = TRUE)
}
\references{
Balassa, B. (1965) Trade Liberalization and Revealed Comparative Advantage, \emph{The Manchester School} \strong{33}: 99-123.
}
\seealso{
\code{\link{location_quotient}}
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\keyword{specialization}
