% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\name{diversity}
\alias{diversity}
\title{Compute a simple measure of diversity of regions}
\usage{
diversity(mat, RCA = FALSE)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns}

\item{RCA}{Logical; should the index of relative comparative advantage (RCA - also refered to as location quotient) first be computed? Defaults to FALSE (a binary matrix - 0/1 - is expected as an input), but can be set to TRUE if the index of relative comparative advantage first needs to be computed}
}
\description{
This function computes a simple measure of diversity of regions by counting the number of industries in which a region has a relative comparative advantage (location quotient > 1) from regions - industries (incidence) matrices
}
\examples{
## generate a region - industry matrix with full count
set.seed(31)
mat <- matrix(sample(0:10,20,replace=T), ncol = 4)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function
diversity (mat, RCA = TRUE)

## generate a region - industry matrix in which cells represent the presence/absence of a RCA
set.seed(31)
mat <- matrix(sample(0:1,20,replace=T), ncol = 4)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function
diversity (mat)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Balland, P.A. and Rigby, D. (2016) The geography of complex knowledge, \emph{Economic Geography, forthcoming}
}
\seealso{
\code{\link{ubiquity}}, \code{\link{location.quotient}}
}
\keyword{diversity}

