% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.list.r
\name{get.list}
\alias{get.list}
\title{Create regular data frames from regions - industries matrices}
\usage{
get.list (data)
}
\arguments{
\item{mat}{An incidence matrix with regions in rows and industries in columns (or the other way around)}

\item{sparse}{Logical; is the input a sparse matrix? Defaults to FALSE}
}
\description{
This function creates regular data frames with three columns (regions, industries, count) from (incidence) matrices (wide to long format) using the reshape2 package
}
\examples{
## generate a region - industry matrix
set.seed(31)
mat <- matrix(sample(0:100,20,replace=T), ncol = 4)
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function
get.list (mat)

## generate a sparse matrix
mat = Matrix (matrix(sample(0:100,20,replace=T), ncol = 4))
rownames(mat) <- c ("R1", "R2", "R3", "R4", "R5")
colnames(mat) <- c ("I1", "I2", "I3", "I4")

## run the function
get.list (mat, sparse = T)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\seealso{
\code{\link{get.matrix}}
}
\keyword{data.management}

