% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.r
\name{Gini}
\alias{Gini}
\title{Compute the Gini coefficient from regional industrial counts}
\usage{
Gini(ind)
}
\arguments{
\item{ind}{A vector of industrial regional count}
}
\description{
This function computes the Gini coefficient from regional industrial count. This index gives an indication of the unequal distribution of an industry accross regions.
}
\examples{
## generate vectors of industrial count
ind <- c(0, 10, 10, 30, 50)

## run the function
Gini (ind)
}
\author{
Pierre-Alexandre Balland \email{p.balland@uu.nl}
}
\references{
Gini, C. (1921) Measurement of Inequality of Incomes, \emph{The Economic Journal} \strong{31}: 124-126
}
\seealso{
\code{\link{Hoover.curve}}
}
\keyword{Gini}
\keyword{advantage,}
\keyword{comparative}
\keyword{location}
\keyword{quotient,}
\keyword{relative}

